Move information pages to folder
