+++ pages/information/speaker-information.js
@@ -1,8 +1,8 @@
-import Layout from '../components/Layout'
-import Header from "../components/Header"
-import theme from "../theme/theme"
+import Layout from '../../components/Layout'
+import Header from "../../components/Header"
+import theme from "../../theme/theme"
 import Head from "next/head"
-import FullWidthImage from "../components/FullWidthImage"
+import FullWidthImage from "../../components/FullWidthImage"
 
 
 export default () => (
